Change ID types from i32 to i64 (#38)
* Change ID types from i32 to u64 in repository traits

* Change ID types from i32 to u64 in mock implementations

* Change ID types from i32 to u64 in actions

* Change ID types from i32 to u64 in API types and JWT module

* Change ID types from i32 to u64 in session module

* Change ID types from i32 to u64 in PostgreSQL implementations

* Change ID types from i32 to u64 in SQLite core implementations

* Change ID types from i32 to u64 in SQLite teams implementations

* Change ID types from i32 to u64 in teams module

* Change ID types from i32 to u64 in PostgreSQL teams implementations

* Change ID types from i32 to u64 in API handlers

* Add clippy allows for i64/u64 conversions in database modules

* Add missing SQLite example definitions

* Fix feature-gated doctest by marking as ignore

* Update PostgreSQL migrations to use BIGINT for IDs

* Change all ID types from u64 to i64

* Fix clippy warnings from i64 migration

* Unify mock prune_expired casts with allow approach

+++ src/actions/delete_user.rs
@@ -13,7 +13,7 @@ impl<U: UserRepository> DeleteUserAction<U> {
         feature = "tracing",
         tracing::instrument(name = "delete_user", skip_all, err)
     )]
-    pub async fn execute(&self, user_id: i32) -> Result<(), AuthError> {
+    pub async fn execute(&self, user_id: i64) -> Result<(), AuthError> {
         self.user_repository.delete_user(user_id).await?;
 
         log::info!(
