+++ src/postgres/magic_link.rs
@@ -18,7 +18,7 @@ impl PostgresMagicLinkRepository {
 
 #[derive(FromRow)]
 struct MagicLinkTokenRecord {
-    user_id: i32,
+    user_id: i64,
     expires_at: DateTime<Utc>,
     created_at: DateTime<Utc>,
 }
@@ -28,7 +28,7 @@ impl MagicLinkRepository for PostgresMagicLinkRepository {
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
     async fn create_magic_link_token(
         &self,
-        user_id: i32,
+        user_id: i64,
         expires_at: DateTime<Utc>,
     ) -> Result<MagicLinkToken, AuthError> {
         let plain_token = generate_token_default();
@@ -98,7 +98,7 @@ impl MagicLinkRepository for PostgresMagicLinkRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn prune_expired(&self) -> Result<u64, AuthError> {
+    async fn prune_expired(&self) -> Result<i64, AuthError> {
         let result = sqlx::query("DELETE FROM magic_link_tokens WHERE expires_at < NOW()")
             .execute(&self.pool)
             .await
@@ -107,6 +107,6 @@ impl MagicLinkRepository for PostgresMagicLinkRepository {
                 AuthError::DatabaseError(e.to_string())
             })?;
 
-        Ok(result.rows_affected())
+        Ok(result.rows_affected() as i64)
     }
 }
