+++ src/postgres/teams/membership.rs
@@ -18,9 +18,9 @@ impl PostgresTeamMembershipRepository {
 
 #[derive(FromRow)]
 struct MembershipRecord {
-    id: i32,
-    team_id: i32,
-    user_id: i32,
+    id: i64,
+    team_id: i64,
+    user_id: i64,
     role: String,
     created_at: DateTime<Utc>,
     updated_at: DateTime<Utc>,
@@ -64,7 +64,7 @@ impl TeamMembershipRepository for PostgresTeamMembershipRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn find_by_id(&self, id: i32) -> Result<Option<TeamMembership>, AuthError> {
+    async fn find_by_id(&self, id: i64) -> Result<Option<TeamMembership>, AuthError> {
         let row: Option<MembershipRecord> = sqlx::query_as(
             "SELECT id, team_id, user_id, role, created_at, updated_at FROM team_memberships WHERE id = $1",
         )
@@ -82,8 +82,8 @@ impl TeamMembershipRepository for PostgresTeamMembershipRepository {
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
     async fn find_by_team_and_user(
         &self,
-        team_id: i32,
-        user_id: i32,
+        team_id: i64,
+        user_id: i64,
     ) -> Result<Option<TeamMembership>, AuthError> {
         let row: Option<MembershipRecord> = sqlx::query_as(
             "SELECT id, team_id, user_id, role, created_at, updated_at FROM team_memberships WHERE team_id = $1 AND user_id = $2",
@@ -101,7 +101,7 @@ impl TeamMembershipRepository for PostgresTeamMembershipRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn find_by_team(&self, team_id: i32) -> Result<Vec<TeamMembership>, AuthError> {
+    async fn find_by_team(&self, team_id: i64) -> Result<Vec<TeamMembership>, AuthError> {
         let rows: Vec<MembershipRecord> = sqlx::query_as(
             "SELECT id, team_id, user_id, role, created_at, updated_at FROM team_memberships WHERE team_id = $1 ORDER BY created_at ASC",
         )
@@ -117,7 +117,7 @@ impl TeamMembershipRepository for PostgresTeamMembershipRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn find_by_user(&self, user_id: i32) -> Result<Vec<TeamMembership>, AuthError> {
+    async fn find_by_user(&self, user_id: i64) -> Result<Vec<TeamMembership>, AuthError> {
         let rows: Vec<MembershipRecord> = sqlx::query_as(
             "SELECT id, team_id, user_id, role, created_at, updated_at FROM team_memberships WHERE user_id = $1 ORDER BY created_at ASC",
         )
@@ -133,7 +133,7 @@ impl TeamMembershipRepository for PostgresTeamMembershipRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn update_role(&self, id: i32, role: &str) -> Result<TeamMembership, AuthError> {
+    async fn update_role(&self, id: i64, role: &str) -> Result<TeamMembership, AuthError> {
         let row: MembershipRecord = sqlx::query_as(
             r"
             UPDATE team_memberships SET role = $1, updated_at = NOW()
@@ -157,7 +157,7 @@ impl TeamMembershipRepository for PostgresTeamMembershipRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn delete(&self, id: i32) -> Result<(), AuthError> {
+    async fn delete(&self, id: i64) -> Result<(), AuthError> {
         sqlx::query("DELETE FROM team_memberships WHERE id = $1")
             .bind(id)
             .execute(&self.pool)
@@ -171,7 +171,7 @@ impl TeamMembershipRepository for PostgresTeamMembershipRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn delete_by_team_and_user(&self, team_id: i32, user_id: i32) -> Result<(), AuthError> {
+    async fn delete_by_team_and_user(&self, team_id: i64, user_id: i64) -> Result<(), AuthError> {
         sqlx::query("DELETE FROM team_memberships WHERE team_id = $1 AND user_id = $2")
             .bind(team_id)
             .bind(user_id)
