+++ src/postgres/teams/permission.rs
@@ -43,8 +43,8 @@ where
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
     async fn get_permissions(
         &self,
-        team_id: i32,
-        user_id: i32,
+        team_id: i64,
+        user_id: i64,
     ) -> Result<PermissionSet<R, A>, AuthError> {
         let row: Option<PermissionRecord> = sqlx::query_as(
             "SELECT permissions FROM team_member_permissions WHERE team_id = $1 AND user_id = $2",
@@ -73,8 +73,8 @@ where
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self, permissions), err))]
     async fn set_permissions(
         &self,
-        team_id: i32,
-        user_id: i32,
+        team_id: i64,
+        user_id: i64,
         permissions: &PermissionSet<R, A>,
     ) -> Result<(), AuthError> {
         let json_str = permissions.to_json();
@@ -110,8 +110,8 @@ where
     )]
     async fn grant_permission(
         &self,
-        team_id: i32,
-        user_id: i32,
+        team_id: i64,
+        user_id: i64,
         resource: R,
         action: A,
     ) -> Result<(), AuthError> {
@@ -126,8 +126,8 @@ where
     )]
     async fn revoke_permission(
         &self,
-        team_id: i32,
-        user_id: i32,
+        team_id: i64,
+        user_id: i64,
         resource: &R,
         action: &A,
     ) -> Result<(), AuthError> {
@@ -142,8 +142,8 @@ where
     )]
     async fn has_permission(
         &self,
-        team_id: i32,
-        user_id: i32,
+        team_id: i64,
+        user_id: i64,
         resource: &R,
         action: &A,
     ) -> Result<bool, AuthError> {
